debug

+++ main_benchmark_graphregu.py
@@ -382,7 +382,7 @@ if __name__ == "__main__":
     adj, edgeList = generateAdj(zOut, graphType=args.prunetype, para = args.knn_distance+':'+str(args.k), outAdjTag = (args.useGAEembedding or args.useBothembedding))
     adjdense = sp.csr_matrix.todense(adj)
     adjsample = torch.from_numpy(adjdense)
-    adjsample = adjsample.long()
+    adjsample = adjsample.float()
     print("---Pruning takes %s seconds ---" % (time.time() - prune_time))
     if args.saveFlag:
         reconOut = recon.detach().cpu().numpy()
@@ -571,7 +571,7 @@ if __name__ == "__main__":
         adj, edgeList = generateAdj(zOut, graphType=args.prunetype, para = args.knn_distance+':'+str(args.k), outAdjTag = (args.useGAEembedding or args.useBothembedding)) 
         adjdense = sp.csr_matrix.todense(adj)
         adjsample = torch.from_numpy(adjdense)
-        adjsample = adjsample.long()
+        adjsample = adjsample.float()
         print("---Pruning takes %s seconds ---" % (time.time() - prune_time))
 
         # Whether use GAE embedding
